feat(consensus): fix pr
